requirements.txt
aiohttp>=3.8.0
requests>=2.28.0
python-dotenv>=0.21.0
psycopg2-binary>=2.9.0
sqlalchemy>=1.4.0
scikit-learn>=1.0.0
pandas>=1.5.0
numpy>=1.24.0